+++ tests/EllipseTest.php
@@ -1,0 +1,569 @@
+<?php
+
+declare(strict_types=1);
+
+namespace UIAwesome\Html\Svg\Tests;
+
+use PHPUnit\Framework\Attributes\Group;
+use PHPUnit\Framework\TestCase;
+use UIAwesome\Html\Core\Factory\SimpleFactory;
+use UIAwesome\Html\Core\Values\{Aria, DataProperty, Language, Role};
+use UIAwesome\Html\Svg\Ellipse;
+use UIAwesome\Html\Svg\Tests\Support\Stub\DefaultProvider;
+use UIAwesome\Html\Svg\Tests\Support\TestSupport;
+use UIAwesome\Html\Svg\Values\{FillRule, StrokeLineCap, StrokeLineJoin};
+
+/**
+ * Test suite for {@see Ellipse} element functionality and behavior.
+ *
+ * Validates the management and rendering of the SVG `<ellipse>` element according to the SVG 2 and HTML Living Standard
+ * specifications.
+ *
+ * Ensures correct handling, immutability, and validation of the `Ellipse` tag rendering, supporting all global HTML and
+ * SVG 2 attributes, content, and provider-based configuration.
+ *
+ * Test coverage:
+ * - Accurate rendering of the `<ellipse>` element.
+ * - Correct application of global HTML attributes and SVG-specific attributes.
+ * - Error handling for invalid attributes or configuration.
+ * - Immutability of the API, ensuring that setting attributes returns a new instance.
+ * - Integration with configuration providers and global factory defaults.
+ * - Precedence of user-defined attributes over global defaults and provider settings.
+ *
+ * {@see Ellipse} for element implementation details.
+ * {@see SimpleFactory} for default configuration management.
+ * {@see TestSupport} for assertion utilities.
+ *
+ * @copyright Copyright (C) 2025 Terabytesoftw.
+ * @license https://opensource.org/license/bsd-3-clause BSD 3-Clause License.
+ */
+#[Group('svg')]
+final class EllipseTest extends TestCase
+{
+    use TestSupport;
+
+    public function testRenderWithAddAriaAttribute(): void
+    {
+        self::equalsWithoutLE(
+            <<<HTML
+            <ellipse aria-pressed="true">
+            HTML,
+            Ellipse::tag()->addAriaAttribute('pressed', true)->render(),
+            "Failed asserting that element renders correctly with 'addAriaAttribute()' method.",
+        );
+    }
+
+    public function testRenderWithAddAriaAttributeUsingEnum(): void
+    {
+        self::equalsWithoutLE(
+            <<<HTML
+            <ellipse aria-pressed="true">
+            HTML,
+            Ellipse::tag()->addAriaAttribute(Aria::PRESSED, true)->render(),
+            "Failed asserting that element renders correctly with 'addAriaAttribute()' method.",
+        );
+    }
+
+    public function testRenderWithAddDataAttribute(): void
+    {
+        self::equalsWithoutLE(
+            <<<HTML
+            <ellipse data-value="value">
+            HTML,
+            Ellipse::tag()->addDataAttribute('value', 'value')->render(),
+            "Failed asserting that element renders correctly with 'addDataAttribute()' method.",
+        );
+    }
+
+    public function testRenderWithAddDataAttributeUsingEnum(): void
+    {
+        self::equalsWithoutLE(
+            <<<HTML
+            <ellipse data-value="value">
+            HTML,
+            Ellipse::tag()->addDataAttribute(DataProperty::VALUE, 'value')->render(),
+            "Failed asserting that element renders correctly with 'addDataAttribute()' method.",
+        );
+    }
+
+    public function testRenderWithAriaAttributes(): void
+    {
+        self::equalsWithoutLE(
+            <<<HTML
+            <ellipse aria-controls="modal-1" aria-hidden="false" aria-label="Close">
+            HTML,
+            Ellipse::tag()
+                ->ariaAttributes(
+                    [
+                        'controls' => static fn(): string => 'modal-1',
+                        'hidden' => false,
+                        'label' => 'Close',
+                    ],
+                )
+                ->render(),
+            "Failed asserting that element renders correctly with 'ariaAttributes()' method.",
+        );
+    }
+
+    public function testRenderWithAttributes(): void
+    {
+        self::equalsWithoutLE(
+            <<<HTML
+            <ellipse class="value">
+            HTML,
+            Ellipse::tag()->attributes(['class' => 'value'])->render(),
+            "Failed asserting that element renders correctly with 'attributes()' method.",
+        );
+    }
+
+    public function testRenderWithClass(): void
+    {
+        self::equalsWithoutLE(
+            <<<HTML
+            <ellipse class="value">
+            HTML,
+            Ellipse::tag()->class('value')->render(),
+            "Failed asserting that element renders correctly with 'class' attribute.",
+        );
+    }
+
+    public function testRenderWithCx(): void
+    {
+        self::equalsWithoutLE(
+            <<<HTML
+            <ellipse cx="100">
+            HTML,
+            Ellipse::tag()->cx(100)->render(),
+            "Failed asserting that element renders correctly with 'cx' attribute.",
+        );
+    }
+
+    public function testRenderWithCy(): void
+    {
+        self::equalsWithoutLE(
+            <<<HTML
+            <ellipse cy="150">
+            HTML,
+            Ellipse::tag()->cy(150)->render(),
+            "Failed asserting that element renders correctly with 'cy' attribute.",
+        );
+    }
+
+    public function testRenderWithDataAttributes(): void
+    {
+        self::equalsWithoutLE(
+            <<<HTML
+            <ellipse data-value="test-value">
+            HTML,
+            Ellipse::tag()->dataAttributes(['value' => 'test-value'])->render(),
+            "Failed asserting that element renders correctly with 'dataAttributes()' method.",
+        );
+    }
+
+    public function testRenderWithDefaultConfigurationValues(): void
+    {
+        self::equalsWithoutLE(
+            <<<HTML
+            <ellipse class="default-class">
+            HTML,
+            Ellipse::tag(['class' => 'default-class'])->render(),
+            'Failed asserting that default configuration values are applied correctly.',
+        );
+    }
+
+    public function testRenderWithDefaultProvider(): void
+    {
+        self::equalsWithoutLE(
+            <<<HTML
+            <ellipse class="default-class" title="default-title">
+            HTML,
+            Ellipse::tag()->addDefaultProvider(DefaultProvider::class)->render(),
+            'Failed asserting that default provider is applied correctly.',
+        );
+    }
+
+    public function testRenderWithFill(): void
+    {
+        self::equalsWithoutLE(
+            <<<HTML
+            <ellipse fill="#ff0000">
+            HTML,
+            Ellipse::tag()->fill('#ff0000')->render(),
+            "Failed asserting that element renders correctly with 'fill' attribute.",
+        );
+    }
+
+    public function testRenderWithFillOpacity(): void
+    {
+        self::equalsWithoutLE(
+            <<<HTML
+            <ellipse fill-opacity="0.7">
+            HTML,
+            Ellipse::tag()->fillOpacity('0.7')->render(),
+            "Failed asserting that element renders correctly with 'fill-opacity' attribute.",
+        );
+    }
+
+    public function testRenderWithFillRule(): void
+    {
+        self::equalsWithoutLE(
+            <<<HTML
+            <ellipse fill-rule="evenodd">
+            HTML,
+            Ellipse::tag()->fillRule('evenodd')->render(),
+            "Failed asserting that element renders correctly with 'fill-rule' attribute.",
+        );
+    }
+
+    public function testRenderWithFillRuleUsingEnum(): void
+    {
+        self::equalsWithoutLE(
+            <<<HTML
+            <ellipse fill-rule="nonzero">
+            HTML,
+            Ellipse::tag()->fillRule(FillRule::NONZERO)->render(),
+            "Failed asserting that element renders correctly with 'fill-rule' attribute.",
+        );
+    }
+
+    public function testRenderWithGlobalDefaultsAreApplied(): void
+    {
+        SimpleFactory::setDefaults(Ellipse::class, ['class' => 'default-class']);
+
+        self::equalsWithoutLE(
+            <<<HTML
+            <ellipse class="default-class">
+            HTML,
+            Ellipse::tag()->render(),
+            'Failed asserting that global defaults are applied correctly.',
+        );
+
+        SimpleFactory::setDefaults(Ellipse::class, []);
+    }
+
+    public function testRenderWithId(): void
+    {
+        self::equalsWithoutLE(
+            <<<HTML
+            <ellipse id="test-id">
+            HTML,
+            Ellipse::tag()->id('test-id')->render(),
+            "Failed asserting that element renders correctly with 'id' attribute.",
+        );
+    }
+
+    public function testRenderWithLangUsingEnum(): void
+    {
+        self::equalsWithoutLE(
+            <<<HTML
+            <ellipse lang="es">
+            HTML,
+            Ellipse::tag()->lang(Language::SPANISH)->render(),
+            "Failed asserting that element renders correctly with 'lang' attribute.",
+        );
+    }
+
+    public function testRenderWithOpacity(): void
+    {
+        self::equalsWithoutLE(
+            <<<HTML
+            <ellipse opacity="0.5">
+            HTML,
+            Ellipse::tag()->opacity('0.5')->render(),
+            "Failed asserting that element renders correctly with 'opacity' attribute.",
+        );
+    }
+
+    public function testRenderWithPathLength(): void
+    {
+        self::equalsWithoutLE(
+            <<<HTML
+            <ellipse pathLength="200">
+            HTML,
+            Ellipse::tag()->pathLength(200)->render(),
+            "Failed asserting that element renders correctly with 'pathLength' attribute.",
+        );
+    }
+
+    public function testRenderWithRole(): void
+    {
+        self::equalsWithoutLE(
+            <<<HTML
+            <ellipse role="banner">
+            HTML,
+            Ellipse::tag()->role('banner')->render(),
+            "Failed asserting that element renders correctly with 'role' attribute.",
+        );
+    }
+
+    public function testRenderWithRoleUsingEnum(): void
+    {
+        self::equalsWithoutLE(
+            <<<HTML
+            <ellipse role="banner">
+            HTML,
+            Ellipse::tag()->role(Role::BANNER)->render(),
+            "Failed asserting that element renders correctly with 'role' attribute.",
+        );
+    }
+
+    public function testRenderWithRx(): void
+    {
+        self::equalsWithoutLE(
+            <<<HTML
+            <ellipse rx="80">
+            HTML,
+            Ellipse::tag()->rx(80)->render(),
+            "Failed asserting that element renders correctly with 'rx' attribute.",
+        );
+    }
+
+    public function testRenderWithRy(): void
+    {
+        self::equalsWithoutLE(
+            <<<HTML
+            <ellipse ry="40">
+            HTML,
+            Ellipse::tag()->ry(40)->render(),
+            "Failed asserting that element renders correctly with 'ry' attribute.",
+        );
+    }
+
+    public function testRenderWithStroke(): void
+    {
+        self::equalsWithoutLE(
+            <<<HTML
+            <ellipse stroke="#00ff00">
+            HTML,
+            Ellipse::tag()->stroke('#00ff00')->render(),
+            "Failed asserting that element renders correctly with 'stroke' attribute.",
+        );
+    }
+
+    public function testRenderWithStrokeDashArray(): void
+    {
+        self::equalsWithoutLE(
+            <<<HTML
+            <ellipse stroke-dasharray="5,5">
+            HTML,
+            Ellipse::tag()->strokeDashArray('5,5')->render(),
+            "Failed asserting that element renders correctly with 'stroke-dasharray' attribute.",
+        );
+    }
+
+    public function testRenderWithStrokeLineCap(): void
+    {
+        self::equalsWithoutLE(
+            <<<HTML
+            <ellipse stroke-linecap="round">
+            HTML,
+            Ellipse::tag()->strokeLineCap('round')->render(),
+            "Failed asserting that element renders correctly with 'stroke-linecap' attribute.",
+        );
+    }
+
+    public function testRenderWithStrokeLineCapUsingEnum(): void
+    {
+        self::equalsWithoutLE(
+            <<<HTML
+            <ellipse stroke-linecap="square">
+            HTML,
+            Ellipse::tag()->strokeLineCap(StrokeLineCap::SQUARE)->render(),
+            "Failed asserting that element renders correctly with 'stroke-linecap' attribute.",
+        );
+    }
+
+    public function testRenderWithStrokeLineJoin(): void
+    {
+        self::equalsWithoutLE(
+            <<<HTML
+            <ellipse stroke-linejoin="round">
+            HTML,
+            Ellipse::tag()->strokeLineJoin('round')->render(),
+            "Failed asserting that element renders correctly with 'stroke-linejoin' attribute.",
+        );
+    }
+
+    public function testRenderWithStrokeLineJoinUsingEnum(): void
+    {
+        self::equalsWithoutLE(
+            <<<HTML
+            <ellipse stroke-linejoin="round">
+            HTML,
+            Ellipse::tag()->strokeLineJoin(StrokeLineJoin::ROUND)->render(),
+            "Failed asserting that element renders correctly with 'stroke-linejoin' attribute.",
+        );
+    }
+
+    public function testRenderWithStrokeMiterlimit(): void
+    {
+        self::equalsWithoutLE(
+            <<<HTML
+            <ellipse stroke-miterlimit="10">
+            HTML,
+            Ellipse::tag()->strokeMiterlimit(10)->render(),
+            "Failed asserting that element renders correctly with 'stroke-miterlimit' attribute.",
+        );
+    }
+
+    public function testRenderWithStrokeOpacity(): void
+    {
+        self::equalsWithoutLE(
+            <<<HTML
+            <ellipse stroke-opacity="0.8">
+            HTML,
+            Ellipse::tag()->strokeOpacity(0.8)->render(),
+            "Failed asserting that element renders correctly with 'stroke-opacity' attribute.",
+        );
+    }
+
+    public function testRenderWithStrokeWidth(): void
+    {
+        self::equalsWithoutLE(
+            <<<HTML
+            <ellipse stroke-width="2">
+            HTML,
+            Ellipse::tag()->strokeWidth('2')->render(),
+            "Failed asserting that element renders correctly with 'stroke-width' attribute.",
+        );
+    }
+
+    public function testRenderWithStyle(): void
+    {
+        self::equalsWithoutLE(
+            <<<HTML
+            <ellipse style='test-value'>
+            HTML,
+            Ellipse::tag()->style('test-value')->render(),
+            "Failed asserting that element renders correctly with 'style' attribute.",
+        );
+    }
+
+    public function testRenderWithToString(): void
+    {
+        self::equalsWithoutLE(
+            <<<HTML
+            <ellipse>
+            HTML,
+            (string) Ellipse::tag(),
+            "Failed asserting that '__toString()' method renders correctly.",
+        );
+    }
+
+    public function testRenderWithTransform(): void
+    {
+        self::equalsWithoutLE(
+            <<<HTML
+            <ellipse transform="rotate(45)">
+            HTML,
+            Ellipse::tag()->transform('rotate(45)')->render(),
+            "Failed asserting that element renders correctly with 'transform' attribute.",
+        );
+    }
+
+    public function testRenderWithUserOverridesGlobalDefaults(): void
+    {
+        SimpleFactory::setDefaults(Ellipse::class, ['class' => 'from-global', 'id' => 'id-global']);
+
+        self::equalsWithoutLE(
+            <<<HTML
+            <ellipse class="from-global" id="id-user">
+            HTML,
+            Ellipse::tag(['id' => 'id-user'])->render(),
+            'Failed asserting that user-defined attributes override global defaults correctly.',
+        );
+
+        SimpleFactory::setDefaults(Ellipse::class, []);
+    }
+
+    public function testReturnNewInstanceWhenSettingAttribute(): void
+    {
+        $ellipse = Ellipse::tag();
+
+        self::assertNotSame(
+            $ellipse,
+            $ellipse->cx(0),
+            'Should return a new instance when setting the attribute, ensuring immutability.',
+        );
+        self::assertNotSame(
+            $ellipse,
+            $ellipse->cy(0),
+            'Should return a new instance when setting the attribute, ensuring immutability.',
+        );
+        self::assertNotSame(
+            $ellipse,
+            $ellipse->fill(''),
+            'Should return a new instance when setting the attribute, ensuring immutability.',
+        );
+        self::assertNotSame(
+            $ellipse,
+            $ellipse->fillOpacity('0'),
+            'Should return a new instance when setting the attribute, ensuring immutability.',
+        );
+        self::assertNotSame(
+            $ellipse,
+            $ellipse->fillRule(''),
+            'Should return a new instance when setting the attribute, ensuring immutability.',
+        );
+        self::assertNotSame(
+            $ellipse,
+            $ellipse->opacity('0'),
+            'Should return a new instance when setting the attribute, ensuring immutability.',
+        );
+        self::assertNotSame(
+            $ellipse,
+            $ellipse->pathLength(0),
+            'Should return a new instance when setting the attribute, ensuring immutability.',
+        );
+        self::assertNotSame(
+            $ellipse,
+            $ellipse->rx(0),
+            'Should return a new instance when setting the attribute, ensuring immutability.',
+        );
+        self::assertNotSame(
+            $ellipse,
+            $ellipse->ry(0),
+            'Should return a new instance when setting the attribute, ensuring immutability.',
+        );
+        self::assertNotSame(
+            $ellipse,
+            $ellipse->stroke(''),
+            'Should return a new instance when setting the attribute, ensuring immutability.',
+        );
+        self::assertNotSame(
+            $ellipse,
+            $ellipse->strokeDashArray(''),
+            'Should return a new instance when setting the attribute, ensuring immutability.',
+        );
+        self::assertNotSame(
+            $ellipse,
+            $ellipse->strokeLineCap(''),
+            'Should return a new instance when setting the attribute, ensuring immutability.',
+        );
+        self::assertNotSame(
+            $ellipse,
+            $ellipse->strokeLineJoin(''),
+            'Should return a new instance when setting the attribute, ensuring immutability.',
+        );
+        self::assertNotSame(
+            $ellipse,
+            $ellipse->strokeMiterlimit('1'),
+            'Should return a new instance when setting the attribute, ensuring immutability.',
+        );
+        self::assertNotSame(
+            $ellipse,
+            $ellipse->strokeOpacity('0'),
+            'Should return a new instance when setting the attribute, ensuring immutability.',
+        );
+        self::assertNotSame(
+            $ellipse,
+            $ellipse->strokeWidth(''),
+            'Should return a new instance when setting the attribute, ensuring immutability.',
+        );
+        self::assertNotSame(
+            $ellipse,
+            $ellipse->transform(''),
+            'Should return a new instance when setting the attribute, ensuring immutability.',
+        );
+    }
+}